Remove empty svn:mergeinfo properties so it won't get in the way during future merges.
The server and the repository are Subversion 1.4.x: they are unable to
understand this property.

Subversion 1.5.x clients usually won't commit it on purpose when the
server and/or the repository are 1.4.x. But they will copy it when doing
WC-to-WC copy because they do not communicate with the server at this
time.


git-svn-id: https://svn.process-one.net/ejabberd/tags/ejabberd-2.0.3@1831 1613c18e-e6f4-0310-972c-c3fad1d2be73